matrix/pyproject.toml
[project]
name = "matrix-platform"
version = "0.1.0"
description = "Matrix notification platform and Claude bridge for vibe-coder-framework"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "mcp[cli]>=1.0",
    "starlette>=0.37",
    "python-json-logger>=2.0",
    "psycopg[binary]>=3.2",
    "pyyaml>=6.0",
]

[project.scripts]
matrix-bridge = "matrix_bridge.main:main"
matrix-api    = "matrix_api.main:main"
matrix-mcp    = "matrix_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src/matrix_common",
    "src/matrix_api",
    "src/matrix_mcp",
    "src/matrix_bridge",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
